requirements.txt
feedparser>=6.0.11
trafilatura>=1.6.5
httpx>=0.27.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
